+++ apps/echo-speaks-zones.groovy
@@ -28,8 +28,8 @@ import java.util.concurrent.Semaphore
 //************************************************
 //*               STATIC VARIABLES               *
 //************************************************
-@Field static final String appVersionFLD  = '4.2.0.4'
-@Field static final String appModifiedFLD = '2022-04-21'
+@Field static final String appVersionFLD  = '4.2.0.6'
+@Field static final String appModifiedFLD = '2022-05-05'
 @Field static final String sNULL          = (String)null
 @Field static final String sBLANK         = ''
 @Field static final String sSPACE         = ' '